Move specs to middleware and add  spec  to check name
