+++ spec/browser/middleware/retry/shared-examples.js
